fix UI bugs and implement items business logic
