+++ src/components/Storage/DesktopStorage/DesktopItemsList/Pagination/PaginationDisplay.tsx
@@ -2,7 +2,7 @@ import { Flex, Select, Icon, Text } from '@chakra-ui/react'
 import Link from 'next/link'
 import { BiArrowToLeft, BiArrowToRight } from 'react-icons/bi'
 import { IoIosArrowBack, IoIosArrowForward } from 'react-icons/io'
-import { PaginationSettings } from '../../../UI/PaginationGeneral/PaginationGeneral'
+import { PaginationSettings } from '../../../../UI/PaginationGeneral/PaginationGeneral'
 
 export const SortingMechanism: React.FC<PaginationSettings> = ({
   handleOnChangeQuantity,
