+++ src/components/Storage/MobileStorage/Sorting/Sorting.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { BsFillGridFill } from 'react-icons/bs'
 import { FaThList } from 'react-icons/fa'
 import { Flex, Select, Text, ButtonGroup, Icon } from '@chakra-ui/react'
-import { sortingType } from '../../../utils/types/frontendGeneral'
+import { sortingType } from '../../../../utils/types/frontendGeneral'
 
 const Sorting: React.FC<{
   setListType: (type: sortingType) => void
