+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import type { NextPage } from 'next'
 import { withPageAuthRequired } from '@auth0/nextjs-auth0'
 import { MainViewProps } from '../utils/types/frontendGeneral'
-import MobileStorage from '../components/MobileStorage/MobileStorage'
+import MobileStorage from '../components/Storage/MobileStorage/MobileStorage'
 import { useMediaQuery } from '@chakra-ui/react'
-import DesktopStorage from '../components/DesktopStorage/DesktopStorage'
+import DesktopStorage from '../components/Storage/DesktopStorage/DesktopStorage'
 import { connectDB } from '../mongo/db'
 import { Text } from '@chakra-ui/react'
 import * as itemsService from '../services/itemsService'
